refactor: Separate core logic from the main library and add new abstractions for easy extension
