--- packages/widgetbook_golden_test/lib/src/create_golden_test.dart
@@ -1,151 +1,0 @@
-import 'dart:async';
-import 'dart:io';
-
-import 'package:flutter/material.dart';
-import 'package:flutter_test/flutter_test.dart';
-import 'package:mocktail/mocktail.dart';
-import 'package:widgetbook/widgetbook.dart';
-import 'package:widgetbook_golden_test_core/src/test_http_overrides.dart';
-import 'package:widgetbook_golden_test/src/widget_tester_extension.dart';
-import 'package:widgetbook_golden_test_core/widgetbook_golden_test_core.dart';
-
-class _BuildContextMock extends Mock implements BuildContext {}
-
-/// Creates and runs a golden test of the given Widgetbook [useCase].
-/// The golden snapshot is saved in [goldenSnapshotsOutputPath]
-void createGoldenTest(
-  WidgetbookUseCase useCase,
-  String goldenSnapshotsOutputPath,
-  WidgetbookGoldenTestsProperties properties,
-) {
-  WidgetbookGoldenTestBuilder? goldenTestBuilder;
-  final mockContext = _BuildContextMock();
-  try {
-    final widget = useCase.build(mockContext);
-    if (widget is WidgetbookGoldenTestBuilder) {
-      goldenTestBuilder = widget;
-    }
-  } catch (e) {
-    // Not a WidgetbookGoldenTestBuilder.
-  }
-
-  // Skip the golden test case if it contains the [skip-golden] tag just as a fallback.
-  bool shouldSkip =
-      (goldenTestBuilder?.skip ?? false) ||
-      // ignore: deprecated_member_use
-      useCase.name.contains(properties.skipTag);
-  // Golden test case of the story.
-  testWidgetsGolden(useCase.name, properties, shouldSkip, (widgetTester) async {
-    final widgetToTest = await widgetTester.pumpWidgetbookCase(
-      properties,
-      useCase,
-      goldenTestBuilder?.addons,
-    );
-
-    await expectLater(
-      find.byType(widgetToTest.runtimeType).first,
-      matchesGoldenFile("$goldenSnapshotsOutputPath/${useCase.name}.png"),
-    );
-  });
-
-  if (goldenTestBuilder?.goldenActions != null) {
-    for (final play in goldenTestBuilder!.goldenActions!) {
-      testWidgetsGolden(
-        "${useCase.name} - ${play.name}",
-        properties,
-        shouldSkip,
-        (widgetTester) async {
-          await validatePlayFunction(
-            widgetTester,
-            properties,
-            useCase,
-            play,
-            goldenSnapshotsOutputPath,
-            goldenTestBuilder?.addons,
-          );
-        },
-      );
-    }
-  }
-}
-
-@visibleForTesting
-void testWidgetsGolden(
-  String description,
-  WidgetbookGoldenTestsProperties properties,
-  bool skip,
-  Future<void> Function(WidgetTester widgetTester) testBody,
-) {
-  testWidgets(description, (widgetTester) async {
-    final previousOnError = FlutterError.onError;
-    await runZonedGuarded<Future<void>>(
-      () async {
-        await HttpOverrides.runZoned(() async {
-          FlutterError.onError = (FlutterErrorDetails details) async {
-            // Ignore image loading errors for generated by the mocked http client.
-            handleError(details, properties, previousOnError);
-          };
-          try {
-            await testBody(widgetTester);
-          } finally {
-            FlutterError.onError = previousOnError;
-          }
-        }, createHttpClient: (_) => createHttpClient(properties));
-      },
-      (error, stack) {
-        var details = FlutterErrorDetails(
-          exception: error,
-          stack: stack,
-          context: ErrorDescription(error.toString()),
-        );
-        handleError(details, properties, null);
-      },
-    );
-  }, skip: skip);
-}
-
-Future<void> validatePlayFunction(
-  WidgetTester widgetTester,
-  WidgetbookGoldenTestsProperties properties,
-  WidgetbookUseCase useCase,
-  GoldenPlayAction play,
-  String goldenSnapshotsOutputPath,
-  List<WidgetbookAddon>? builderAddons,
-) async {
-  final widgetToTest = await widgetTester.pumpWidgetbookCase(
-    properties,
-    useCase,
-    builderAddons,
-  );
-  await play.callback(widgetTester, find);
-  await widgetTester.pumpAndSettle();
-  Finder goldenFinder = play.goldenFinder == null
-      ? find.byType(widgetToTest.runtimeType).first
-      : play.goldenFinder!.call(find);
-  await expectLater(
-    goldenFinder,
-    matchesGoldenFile(
-      "$goldenSnapshotsOutputPath/${useCase.name} - ${play.name}.png",
-    ),
-  );
-}
-
-@visibleForTesting
-void handleError(
-  FlutterErrorDetails details,
-  WidgetbookGoldenTestsProperties properties,
-  Function(FlutterErrorDetails)? previousOnError,
-) {
-  // Ignore image loading errors for generated by the mocked http client.
-  if (details.exception is IgnoreNetworkImageException) {
-    return;
-  }
-
-  if (properties.onTestError != null) {
-    properties.onTestError!(details, previousOnError);
-  } else if (previousOnError != null) {
-    previousOnError(details);
-  } else {
-    throw details.exception;
-  }
-}
