+++ packages/widgetbook_golden_test/lib/widgetbook_golden_test.dart
@@ -1,2 +1,1 @@
 export 'src/run_widgetbook_golden_tests.dart';
-export 'src/create_golden_test.dart';
