+++ packages/widgetbook_golden_test_core/lib/src/helpers/golden_test_zone_runner.dart
@@ -1,0 +1,56 @@
+import 'dart:async';
+import 'dart:io';
+
+import 'package:flutter/material.dart';
+import 'package:widgetbook_golden_test_core/src/helpers/test_http_overrides.dart';
+import 'package:widgetbook_golden_test_core/widgetbook_golden_test_core.dart';
+
+Future<void> goldenTestZoneRunner({
+  required Future<void> Function() testBody,
+  required WidgetbookGoldenTestsProperties properties,
+}) async {
+  final previousOnError = FlutterError.onError;
+  await runZonedGuarded<Future<void>>(
+    () async {
+      await HttpOverrides.runZoned(() async {
+        FlutterError.onError = (FlutterErrorDetails details) async {
+          // Ignore image loading errors for generated by the mocked http client.
+          handleError(details, properties, previousOnError);
+        };
+        try {
+          await testBody();
+        } finally {
+          FlutterError.onError = previousOnError;
+        }
+      }, createHttpClient: (_) => createHttpClient(properties));
+    },
+    (error, stack) {
+      var details = FlutterErrorDetails(
+        exception: error,
+        stack: stack,
+        context: ErrorDescription(error.toString()),
+      );
+      handleError(details, properties, null);
+    },
+  );
+}
+
+@visibleForTesting
+void handleError(
+  FlutterErrorDetails details,
+  WidgetbookGoldenTestsProperties properties,
+  Function(FlutterErrorDetails)? previousOnError,
+) {
+  // Ignore image loading errors for generated by the mocked http client.
+  if (details.exception is IgnoreNetworkImageException) {
+    return;
+  }
+
+  if (properties.onTestError != null) {
+    properties.onTestError!(details, previousOnError);
+  } else if (previousOnError != null) {
+    previousOnError(details);
+  } else {
+    throw details.exception;
+  }
+}
