+++ packages/widgetbook_golden_test_core/lib/src/helpers/ignore_network_image_exception.dart
