+++ packages/widgetbook_golden_test_core/lib/src/widgetbook_golden_test_renderer.dart
@@ -1,0 +1,21 @@
+import 'package:widgetbook/widgetbook.dart';
+import 'package:widgetbook_golden_test_core/widgetbook_golden_test_core.dart';
+
+abstract class WidgetbookGoldenRenderer {
+  void renderSimpleGoldenTest({
+    required WidgetbookUseCase useCase,
+    required String goldenPath,
+    required WidgetbookGoldenTestsProperties properties,
+    required bool skip,
+    required WidgetbookGoldenTestBuilder? goldenTestBuilder,
+  });
+
+  void renderGoldenPlayActionTest({
+    required WidgetbookUseCase useCase,
+    required String goldenPath,
+    required WidgetbookGoldenTestsProperties properties,
+    required GoldenPlayAction action,
+    required bool skip,
+    required WidgetbookGoldenTestBuilder? goldenTestBuilder,
+  });
+}
