+++ packages/widgetbook_golden_test_core/lib/widgetbook_golden_test_core.dart
@@ -1,4 +1,7 @@
+export 'src/golden_play_action.dart';
 export 'src/widgetbook_golden_tests_properties.dart';
 export 'src/widgetbook_golden_test_builder.dart';
-export 'src/golden_play_action.dart';
-export 'src/ignore_network_image_exception.dart';
+export 'src/widgetbook_golden_test_generator.dart';
+export 'src/widgetbook_golden_test_renderer.dart';
+export 'src/helpers/ignore_network_image_exception.dart';
+export 'src/helpers/golden_test_zone_runner.dart';
