+++ packages/widgetbook_golden_test_core/pubspec.yaml
@@ -20,6 +20,7 @@ dependencies:
 
 dev_dependencies:
   flutter_lints: ^5.0.0
+  provider: ^6.1.5+1
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
