+++ packages/widgetbook_golden_test_core/test/unit/helpers/ignore_network_image_exception_test.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart' show expect, test;
-import 'package:widgetbook_golden_test_core/src/ignore_network_image_exception.dart';
+import 'package:widgetbook_golden_test_core/src/helpers/ignore_network_image_exception.dart';
 
 void main() {
   test("toString() returns custom message", () {
